pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "daq_system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "numpy>=1.21.0",
    "openpyxl>=3.0.0",
    "pandas>=1.5.0",
    "pyyaml>=6.0.2",
    "synnax>=0.39.0",
]

[tool.setuptools]
packages = ["daq_system"]
